+++ src/datasources/transaction-api/transaction-api.manager.ts
@@ -4,8 +4,6 @@ import { TransactionApi } from './transaction-api.service';
 import { CacheFirstDataSource } from '../cache/cache.first.data.source';
 import { ITransactionApiManager } from '../../domain/interfaces/transaction-api.manager.interface';
 import { IConfigApi } from '../../domain/interfaces/config-api.interface';
-import { ValidationErrorFactory } from '../errors/validation-error-factory';
-import { JsonSchemaService } from '../../common/schemas/json-schema.service';
 
 @Injectable()
 export class TransactionApiManager implements ITransactionApiManager {
@@ -15,8 +13,6 @@ export class TransactionApiManager implements ITransactionApiManager {
   constructor(
     @Inject(IConfigApi) private readonly configApi: IConfigApi,
     private readonly dataSource: CacheFirstDataSource,
-    private readonly validationErrorFactory: ValidationErrorFactory,
-    private readonly jsonSchemaService: JsonSchemaService,
   ) {}
 
   async getTransactionApi(chainId: string): Promise<TransactionApi> {
@@ -32,8 +28,6 @@ export class TransactionApiManager implements ITransactionApiManager {
       chainId,
       chain.transactionService,
       this.dataSource,
-      this.validationErrorFactory,
-      this.jsonSchemaService,
     );
     return this.transactionApiMap[chainId];
   }
