+++ src/domain/balances/balances.repository.ts
@@ -2,13 +2,29 @@ import { IBalancesRepository } from './balances.repository.interface';
 import { Balance } from './entities/balance.entity';
 import { Inject, Injectable } from '@nestjs/common';
 import { ITransactionApiManager } from '../interfaces/transaction-api.manager.interface';
+import { DefinedError, ValidateFunction } from 'ajv';
+import { JsonSchemaService } from '../../common/schemas/json-schema.service';
+import { ValidationErrorFactory } from '../../datasources/errors/validation-error-factory';
+import {
+  balanceTokenSchema,
+  balanceSchema,
+} from './entities/schemas/balance.schema';
 
 @Injectable()
 export class BalancesRepository implements IBalancesRepository {
+  private readonly isValidBalance: ValidateFunction<Balance>;
+
   constructor(
     @Inject(ITransactionApiManager)
     private readonly transactionApiManager: ITransactionApiManager,
-  ) {}
+    private readonly validationErrorFactory: ValidationErrorFactory,
+    private readonly jsonSchemaService: JsonSchemaService,
+  ) {
+    this.jsonSchemaService.addSchema(balanceTokenSchema, 'balanceToken');
+    this.isValidBalance = this.jsonSchemaService.compile(
+      balanceSchema,
+    ) as ValidateFunction<Balance>;
+  }
 
   async getBalances(
     chainId: string,
@@ -17,6 +33,14 @@ export class BalancesRepository implements IBalancesRepository {
     excludeSpam?: boolean,
   ): Promise<Balance[]> {
     const api = await this.transactionApiManager.getTransactionApi(chainId);
-    return api.getBalances(safeAddress, trusted, excludeSpam);
+    const balances = await api.getBalances(safeAddress, trusted, excludeSpam);
+
+    if (!balances.every((balance) => this.isValidBalance(balance))) {
+      // TODO: probably we want to invalidate cache at this point
+      const errors = this.isValidBalance.errors as DefinedError[];
+      throw this.validationErrorFactory.from(errors);
+    }
+
+    return balances;
   }
 }
